chore: rename primitives files to snake_case
Rust convention: words in file names separated by underscores.
