+++ src/primitives/non_negative_int.rs
